+++ psRadiantOne/Public/Tasks/Get-R1Task.ps1
@@ -9,6 +9,7 @@ function Get-R1Task {
 			ParameterSetName = 'Id'
 		)]
 		[ValidateNotNullOrEmpty()]
+		[Alias('taskId')]
 		[string]$id
 	)